requirements.txt
Flask
gunicorn
Werkzeug
google-generativeai
